Modifications and corrections to get a working version
The original script was not working properly on my system (Debian 7.0
"wheezy"). The phylip sub-programs invocation needs to be preceeded by
"phylip", and the "infile" was not formatted properly (missing space
in the phylip format between the sequence name and the sequence
itself).

I've added a fasta file containing a few rRNA V9 sequences extracted
from the PR2 database, and the tree file expected after a normal run
of pairdist.py.